move pages publications to new pages folder
